[Refactor] (test) Move the module '_data' to the root path of package 'test' and modify importing path of others which uses it.
